requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
networkx>=3.0
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.66
torch>=2.2
torch-geometric>=2.5
jupyter>=1.0
